helper function for send email
