+++ src/helpers/sendTenantResponse.js
@@ -1,6 +1,6 @@
 import { createClient } from "contentful-management";
 import { request } from "https";
-import emailjs from "emailjs-com";
+import { sendRequestUpdateEmail } from "./sendRequestUpdateEmail";
 
 var client = createClient({
   accessToken: process.env.REACT_APP_CONTENT_MANAGEMENT_API
@@ -54,7 +54,7 @@ const postTenantResponse = (response, property, request, status) => {
               })
               .then(requestToUpdate => requestToUpdate.publish())
               .then(updatedRequest => {
-                sendEmail(property, response);
+                sendRequestUpdateEmail(property, response);
                 return updatedRequest;
               })
               .catch(error => {
@@ -66,7 +66,6 @@ const postTenantResponse = (response, property, request, status) => {
           .catch(error => console.log("Error creating entry", error));
       })
       .then(newNotification => {
-        sendEmail(property, response);
         return newNotification;
       })
       .catch(e => {
@@ -74,22 +73,6 @@ const postTenantResponse = (response, property, request, status) => {
         return { error: e };
       })
   );
-};
-
-const sendEmail = (property, response) => {
-  const template_params = {
-    reply_to: property.fields.tenant[0].fields.email,
-    to_name: property.fields.manager.fields.name,
-    to_email: property.fields.manager.fields.email,
-    property_name: property.fields.name,
-    message: response.fields.message["en-US"],
-    subject: response.fields.subject["en-US"]
-  };
-
-  const service_id = "default_service";
-  const template_id = "managerResponse";
-  const user_id = "user_MsiQ3UxI8JGshxx5VNpt5";
-  emailjs.send(service_id, template_id, template_params, user_id);
 };
 
 export const sendTenantResponse = (values, property, request, status) => {
